Moving the logic of mounting config maps and secrets to userVolumes.go file
